fix json path

+++ app/create_jsonfile.py
@@ -1,19 +1,14 @@
-import json
 import os
 
 
-jsonfile_name = "/root/fast_api_demo/jsonfile_test.json"
-# jsonfile_name = "jsonfile_test.json"
-
-
-json_str = '{"a": "a"}'
+# jsonfile_name = "/root/fast_api_demo/jsonfile.json"
+jsonfile_name = "./jsonfile.json"
 
 
 def creat_jsonfile():
-    with open(jsonfile_name, 'w') as jsonfile:
-        jsonfile.write(json_str)
-        print(jsonfile)
-        jsonfile.close()
-
-
-creat_jsonfile()
+    if os.path.exists(jsonfile_name):
+        return
+    else:
+        with open(jsonfile_name, 'w') as jsonfile:
+            jsonfile.write()
+            jsonfile.close()
